refreshes client after token recieved in signin

+++ backend/src/resolvers/updateUser.ts
@@ -10,14 +10,12 @@ import {
 import sanitize from 'mongo-sanitize'
 import { AuthenticationError }  from 'apollo-server-express';
 
-
 export const resolvers = {
   Query: {
     squash: async (parents, unSanitizedId, context, info) => {
       const user = context.user;
       const { id } = sanitize(unSanitizedId);
       if (user?.uid != id) throw new AuthenticationError("not logged in");
-      console.log(id);
       const squash_val = await Squash.findById(id);
       console.log(squash_val);
       return squash_val;
@@ -31,7 +29,7 @@ export const resolvers = {
     },
   },
   Mutation: {
-    updateUserProfile: async (root, unSanitizedData) => {
+    updateUserProfile: async (root, unSanitizedData, context) => {
       const {
         _id,
         image_set,
@@ -47,26 +45,24 @@ export const resolvers = {
         original_uploaded_image_set,
       } = sanitize(unSanitizedData);
 
-      //const removed_image_set = await deleteFilesFromGC(remove_uploaded_images, original_uploaded_image_set, add_local_images.length);
+      const user = context.user;
+      if (user?.uid != _id) throw new AuthenticationError("not logged in");
+
       const removed_image_set = await deleteFilesFromAWS(
         remove_uploaded_images,
         original_uploaded_image_set,
         add_local_images.length
       );
-      //const data_set = await creatGCUpload(add_local_images, _id)
       const data_set = await createAWSUpload(add_local_images, _id);
       const final_image_set = removed_image_set.concat(data_set);
-      // if sport changes decrement sportsPerDay
       const check_doc_sports = await Squash.findById(_id);
       var sportChangesPerDay = check_doc_sports?.sportChangesPerDay;
-      // only check if list of sports are the same, user shoul dbe allowed to change gamelevels freely
       const sportsOld = _.map(check_doc_sports?.sports, (sportObj) => {
         return sportObj.sport;
       });
       const sportsNew = _.map(sports, (sportObj) => {
         return sportObj.sport;
       });
-      //if (!_.isEqual(sportsOld, sportsNew)) {
       if (!_.isEqual(check_doc_sports?.sports, sports)) {
         sportChangesPerDay = sportChangesPerDay
           ? sportChangesPerDay - 1
